Clockit/backend/requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.25.0
sqlalchemy>=2.0.0
alembic>=1.13.0
psycopg2-binary>=2.9.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
bcrypt<4.1
pydantic[email-validator]>=2.0.0
python-multipart>=0.0.6